+++ R/TernaryPlot.R
@@ -414,11 +414,11 @@ TernaryPlot <- function (atip = NULL, btip = NULL, ctip = NULL,
 #' 
 #' @importFrom graphics par
 #' @export
-HorizontalGrid <- function (grid.lines = 10, grid.col='grey',
-                            grid.lty='dotted', grid.lwd=par('lwd'),
-                            direction=getOption('ternDirection')) {
+HorizontalGrid <- function (grid.lines = 10, grid.col = 'grey',
+                            grid.lty = 'dotted', grid.lwd = par('lwd'),
+                            direction = getOption('ternDirection', 1L)) {
   
-  if (!(direction %in% 1:4)) stop("Parameter direction must be an integer from 1 to 4")
+  if (!(direction %in% 1:4)) stop("`direction` must be an integer from 1 to 4")
   line_points <- seq(from=0, to=1, length.out=grid.lines + 1L)
   tern_height <- c(sqrt(3/4), 1, sqrt(3/4), 1)[direction]
   tern_width <- c(1, sqrt(3/4), 1, sqrt(3/4), 1)[direction]